requirements.txt
Flask
scikit-learn
pytest
numpy
pylint
flake8
mysql-connector-python
python-dotenv
apache-airflow==2.9.0
apache-airflow-providers-mysql
pandas
streamlit
joblib
